feat: fifth commit

+++ fip_api/src/auth/auth_claims.rs
@@ -1,12 +1,20 @@
 use serde::{Deserialize, Serialize};
 
+/// TODO: documentation
 #[derive(Debug, Deserialize, Serialize)]
 pub struct AuthClaims {
-    pub aud: String, // Optional. Audience
-    pub exp: usize, // Required (validate_exp defaults to true in validation). Expiration time (as UTC timestamp)
-    pub iat: usize, // Optional. Issued at (as UTC timestamp)
-    pub iss: String, // Optional. Issuer
-    pub jti: String, // Optional. jwt id
-    pub nbf: usize, // Optional. Not Before (as UTC timestamp)
-    pub sub: String, // Optional. Subject (whom token refers to)
+    /// Optional. Audience
+    pub aud: String,
+    /// Required (validate_exp defaults to true in validation). Expiration time (as UTC timestamp)
+    pub exp: i64,
+    /// Optional. Issued at (as UTC timestamp)
+    pub iat: i64,
+    /// Optional. Issuer
+    pub iss: String,
+    /// Optional. jwt id
+    pub jti: String,
+    /// Optional. Not Before (as UTC timestamp)
+    pub nbf: i64,
+    /// Optional. Subject (whom token refers to)
+    pub sub: String,
 }
